Add aux author data for empty search results
In the case where a search is performed with an authorId as a filter,
it is possible for the database to return no results even if the author
exists in the database. Therefore, the database is queried a second
time for contact information associated with only the authorIds. The
author information is then used to update the results returned by the
first database query.

Fix #194