modify uuid + version indicies
Since the uuid value to be matched comes in over the API, we're
exceptionally careful about casting it to a UUID in the backend, to
avoid error falures. Instead, we cast the UUID from the table to a
string and compare strings. Since the index as on the UUID field, it
wasn't being used. Change the indicies to be on the textual version
of the UUID and the combined version.